OSC_DSP_StateMachineUpdate() function is partly written.
The semaphore project is suspended because it would need a great
effort and the it would be slow because it would have to enter
into priviledged mode in order to disable all interrupts.
The menu printer is modified: now the menu element name is shorter
while the menu element value is longer with one letter.