Vertical Offsets are mV based now
Small issue with steps. (sometimes it doesn't steps for a button push)
The Cause of this is the property of the integer division. However
in this way the step values are round numbers. So it is a functionality
vs user confort tradeoff.
The offset is rounded if there is any change in the resolution.